bug fix solved on creating, reading, appending and deleting files

+++ src/threads/DelayDelete.java
@@ -2,13 +2,17 @@ package threads;
 
 import java.io.File;
 
+import fileManagement.FileSetting;
+
 public class DelayDelete implements Runnable{
 	File file;
 	long delay;
+	boolean write;
 	
-	public DelayDelete(String _file, long _delay){
+	public DelayDelete(String _file, long _delay, boolean _write){
 		file = new File(_file);
 		delay = _delay;
+		write = _write;
 	}
 
 	@Override
@@ -19,7 +23,12 @@ public class DelayDelete implements Runnable{
 			e.printStackTrace();
 		}
 		if(file.exists()) {
-			file.delete();
+			if(write == false) {
+				file.delete();
+			}
+			else {
+				FileSetting.createFile(file.getPath(), "expired");
+			}
 		}
 	}
 }
